Include utils/Popup in entry and export reference
`src/index.js` is missing an export of the `PopupManager` singleton. This singleton handles zIndex ordering. When you (a) need to change the base zIndex or (b) want to use the `PopupManager` in your own extensions you need a reference to the  `PopupManager` that ElementUI uses. Therefore the `PopupManager` singleton should be exported in the main entry.